requirements.txt
Pillow
wxwidgets~=1.0.5
betterconf~=2.5.0